renamed models to schemas
